Decision/requirements.txt
pandas
numpy
scipy
# PyPortfolioOpt (Optional, handled by scipy fallback)
scikit-learn
google-generativeai # For Gemini API
